Collect_result , verzamel testresultaten

Verzamel de testresultaten uit een data.frame met tests. De testresultaten worden per regel in de data.frame aangegeven.

collect_result(x)

Arguments

x
data.frame dat getest is door 1 of meerdere QC tests

Value

Een data.frame met voor iedere test in x en voor iedere record / regel met gegeven het test-oordeel. Het test-oordeel wordt gegeven als een factor

Details

Deze functie verzameld de testresultaten die door de afzonderlijk tests zijn toegevoegd aan de data.frame met metingen. Voor iedere test wordt het testresultaat opgehaald en en per meting weergegeven. Het resultaat is een data.frame met per kolom de afzonderlijke test en per meting (regel/record) het resultaat.

De waarden in de data.frame zijn opgeslagen als factors. Hierdoor is het eenvoudig om een snel een overzicht te krijgen door bijvoorbeeld het gebruik van sumarise of plot functies.

Examples

 data(veld)
 data(put)
 data(metingen)
 data(filter)
 x <- QC0a(d_veld = veld, d_put = put, d_metingen = metingen, verbose = TRUE)

#> "Er zijn in totaal 10 bemonsterde putlocaties met XY-coordinaten afwijkend van de BRO geregistreerde putcoordinaten."
#>    putcode  X_val  Y_val  X_BRO  Y_BRO
#> 1        8 227450 456217 227450 456260
#> 2       11 242845 441905 242805 441905
#> 3       97 145365 401775 145325 401775
#> 4      100 165225 422607 165225 422650
#> 5      175 215930 551007 215930 551050
#> 6      190 176490 526350 176450 526350
#> 7      192 188865 517850 188825 517850
#> 8      194 190878 507513 190838 507513
#> 9      199 205950 513482 205950 513525
#> 10     280 133725 526970 133725 527013

 d <- veld %>% dplyr::filter(landgebruik != "heide")
 x <- QC0b(d_veld = d, d_put = put, d_metingen = x, verbose = FALSE)
 x <- QC0f(d_veld = veld, d_filter = filter, d_metingen = x)
 x <- QC0g(d_filter = filter, d_metingen = x)
 x <- QC1a(d_parameter = parameter, d_metingen = x)
 x <- QC1b(d_veld = veld, d_parameter = parameter, d_metingen = x)
 x <- QC1e(d_metingen = x)
 oordeel <- collect_result(x)
 summary(oordeel)

#>       qcid                QC0a                   QC0f       
#>  Min.   :     1   onverdacht:210023   onverdacht   :211304  
#>  1st Qu.: 52949   verdacht  :  1769   twijfelachtig:   488  
#>  Median :105896                                             
#>  Mean   :105896                                             
#>  3rd Qu.:158844                                             
#>  Max.   :211792                                             
#>             QC0g                QC1a                QC1b       
#>  onverdacht   :211426   onverdacht:197904   onverdacht: 72384  
#>  twijfelachtig:   366   verdacht  : 13888   verdacht  :139408  
#>                                                                
#>                                                                
#>                                                                
#>                                                                
#>          QC1e       
#>  ontbrekend:211792  
#>                     
#>                     
#>                     
#>                     
#>