Collect_result , verzamel testresultaten

Verzamel de testresultaten uit een data.frame met tests. De testresultaten worden per regel in de data.frame aangegeven.

collect_result(x)

Arguments

x

data.frame dat getest is door 1 of meerdere QC tests

Value

Een data.frame met voor iedere test in x en voor iedere record / regel met gegeven het test-oordeel. Het test-oordeel wordt gegeven als een factor

Details

Deze functie verzameld de testresultaten die door de afzonderlijk tests zijn toegevoegd aan de data.frame met metingen. Voor iedere test wordt het testresultaat opgehaald en en per meting weergegeven. Het resultaat is een data.frame met per kolom de afzonderlijke test en per meting (regel/record) het resultaat.

De waarden in de data.frame zijn opgeslagen als factors. Hierdoor is het eenvoudig om een snel een overzicht te krijgen door bijvoorbeeld het gebruik van sumarise of plot functies.

Examples

 data(veld)
 data(put)
 data(metingen)
 data(filter)
 x <- QC0a(d_veld = veld, d_put = put, d_metingen = metingen, verbose = TRUE)
#> "Er zijn in totaal 10 bemonsterde putlocaties met XY-coordinaten afwijkend van de BRO geregistreerde putcoordinaten."
#>    putcode  X_val  Y_val  X_BRO  Y_BRO
#> 1       21 226390 567500 226350 567500
#> 2       28 212490 546675 212450 546675
#> 3       78  68760 417805  68720 417805
#> 4       93  85200 446927  85200 446970
#> 5       94 133875 407482 133875 407525
#> 6      183 164930 585020 164890 585020
#> 7      183 164890 584977 164890 585020
#> 8      206 233340 500225 233300 500225
#> 9      272 193288 435470 193288 435513
#> 10     283 125350 521257 125350 521300
 d <- veld %>% dplyr::filter(landgebruik != "heide")
 x <- QC0b(d_veld = d, d_put = put, d_metingen = x, verbose = FALSE)
 x <- QC0f(d_veld = veld, d_filter = filter, d_metingen = x)
 x <- QC0g(d_filter = filter, d_metingen = x)
 x <- QC1a(d_parameter = parameter, d_metingen = x)
 x <- QC1b(d_veld = veld, d_parameter = parameter, d_metingen = x)
 x <- QC1e(d_metingen = x)
 oordeel <- collect_result(x)
 summary(oordeel)
#>       qcid                QC0a                   QC0f       
#>  Min.   :     1   onverdacht:210389   onverdacht   :211365  
#>  1st Qu.: 52949   verdacht  :  1403   twijfelachtig:   427  
#>  Median :105896                                             
#>  Mean   :105896                                             
#>  3rd Qu.:158844                                             
#>  Max.   :211792                                             
#>             QC0g                QC1a                QC1b       
#>  onverdacht   :211426   onverdacht:201376   onverdacht: 72848  
#>  twijfelachtig:   366   verdacht  : 10416   verdacht  :138944  
#>                                                                
#>                                                                
#>                                                                
#>                                                                
#>          QC1e       
#>  ontbrekend:211792  
#>                     
#>                     
#>                     
#>                     
#>